pyproject.toml
[tool.poetry]
name = "almond"
version = "1.0.0"
description = "A monitoring discord.py bot, to alert for high cpu and memory usage"
authors = ["Vitaman02"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"
pydantic = "^1.10.2"
aiohttp = "^3.8.3"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
